Fixed critical bug that crashed the site
`sms.py` did not work, because the new officer objects do not have a previous `uid` field that was trying to be accessed. This field is unnecessary, because the document id can be accessed via `document.id`, and it contains the uid for the user object. The file was changed to use this rather than trying to reference a `uid` field.